Closes #5049: Add WebExtensionToolbarFeature
