+++ components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarPresenterTest.kt
@@ -4,7 +4,6 @@
 
 package mozilla.components.feature.toolbar
 
-import android.graphics.Color
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestCoroutineDispatcher
@@ -15,33 +14,24 @@ import mozilla.components.browser.state.action.TabListAction
 import mozilla.components.browser.state.action.TrackingProtectionAction
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.ContentState
-import mozilla.components.browser.state.state.CustomTabSessionState
 import mozilla.components.browser.state.state.SecurityInfoState
 import mozilla.components.browser.state.state.TabSessionState
 import mozilla.components.browser.state.state.TrackingProtectionState
-import mozilla.components.browser.state.state.WebExtensionState
 import mozilla.components.browser.state.state.createCustomTab
 import mozilla.components.browser.state.state.createTab
 import mozilla.components.browser.state.store.BrowserStore
-import mozilla.components.concept.engine.webextension.BrowserAction
 import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.feature.toolbar.internal.URLRenderer
 import mozilla.components.support.test.any
-import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.ext.joinBlocking
 import mozilla.components.support.test.mock
 import org.junit.After
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
 import org.mockito.Mockito.never
 import org.mockito.Mockito.spy
 import org.mockito.Mockito.verify
 import org.mockito.Mockito.verifyNoMoreInteractions
-
-typealias WebExtensionBrowserAction = mozilla.components.concept.engine.webextension.BrowserAction
 
 class ToolbarPresenterTest {
     private val testDispatcher = TestCoroutineDispatcher()
@@ -109,50 +99,6 @@ class ToolbarPresenterTest {
         verify(toolbar).setSearchTerms("")
         verify(toolbar).displayProgress(0)
         verify(toolbar).siteSecure = Toolbar.SiteSecurity.INSECURE
-    }
-
-    @Test
-    fun `render overridden web extension action from browser state`() {
-        val defaultBrowserAction =
-            WebExtensionBrowserAction("default_title", false, mock(), "", "", 0, 0) {}
-        val overriddenBrowserAction =
-            WebExtensionBrowserAction("overridden_title", false, mock(), "", "", 0, 0) {}
-        val toolbar: Toolbar = mock()
-        val extensions: Map<String, WebExtensionState> = mapOf(
-            "id" to WebExtensionState("id", "url", defaultBrowserAction)
-        )
-        val overriddenExtensions: Map<String, WebExtensionState> = mapOf(
-            "id" to WebExtensionState("id", "url", overriddenBrowserAction)
-        )
-        val store = spy(
-            BrowserStore(
-                BrowserState(
-                    tabs = listOf(
-                        createTab(
-                            "https://www.example.org", id = "tab1",
-                            extensions = overriddenExtensions
-                        )
-                    ), selectedTabId = "tab1",
-                    extensions = extensions
-                )
-            )
-        )
-        val toolbarPresenter = spy(ToolbarPresenter(toolbar, store))
-
-        toolbarPresenter.renderer = mock()
-
-        toolbarPresenter.start()
-
-        testDispatcher.advanceUntilIdle()
-
-        verify(store).observeManually(any())
-        verify(toolbarPresenter).render(any())
-
-        val delegateCaptor = argumentCaptor<WebExtensionToolbarAction>()
-
-        verify(toolbarPresenter.renderer).post("https://www.example.org")
-        verify(toolbar).addBrowserAction(delegateCaptor.capture())
-        assertEquals("overridden_title", delegateCaptor.value.browserAction.title)
     }
 
     @Test
@@ -471,57 +417,5 @@ class ToolbarPresenterTest {
         verify(toolbar).setSearchTerms("")
         verify(toolbar).displayProgress(0)
         verify(toolbar).siteSecure = Toolbar.SiteSecurity.INSECURE
-    }
-
-    @Test
-    fun `WebExtensionBrowserAction is replaced when the web extension is already in the toolbar`() {
-        val toolbarPresenter = ToolbarPresenter(mock(), mock())
-
-        val browserAction = BrowserAction(
-            title = "title",
-            icon = mock(),
-            enabled = true,
-            badgeText = "badgeText",
-            badgeTextColor = Color.WHITE,
-            badgeBackgroundColor = Color.BLUE,
-            uri = "uri"
-        ) {}
-
-        val browserActionDisabled = BrowserAction(
-            title = "title",
-            icon = mock(),
-            enabled = false,
-            badgeText = "badgeText",
-            badgeTextColor = Color.WHITE,
-            badgeBackgroundColor = Color.BLUE,
-            uri = "uri"
-        ) {}
-
-        // Verify browser extension toolbar rendering
-        val browserExtensions = HashMap<String, WebExtensionState>()
-        browserExtensions["1"] = WebExtensionState(id = "1", browserAction = browserAction)
-        browserExtensions["2"] = WebExtensionState(id = "2", browserAction = browserActionDisabled)
-
-        val browserState = BrowserState(extensions = browserExtensions)
-        toolbarPresenter.renderWebExtensionActions(browserState, mock())
-
-        assertTrue(toolbarPresenter.webExtensionBrowserActions.size == 2)
-        assertTrue(toolbarPresenter.webExtensionBrowserActions["1"]!!.browserAction.enabled)
-        assertFalse(toolbarPresenter.webExtensionBrowserActions["2"]!!.browserAction.enabled)
-
-        // Verify tab with existing extension in the toolbar to update its BrowserAction
-        val tabExtensions = HashMap<String, WebExtensionState>()
-        tabExtensions["3"] = WebExtensionState(id = "1", browserAction = browserActionDisabled)
-
-        val tabSessionState = CustomTabSessionState(
-            content = mock(),
-            config = mock(),
-            extensionState = tabExtensions
-        )
-        toolbarPresenter.renderWebExtensionActions(browserState, tabSessionState)
-
-        assertTrue(toolbarPresenter.webExtensionBrowserActions.size == 2)
-        assertFalse(toolbarPresenter.webExtensionBrowserActions["1"]!!.browserAction.enabled)
-        assertFalse(toolbarPresenter.webExtensionBrowserActions["2"]!!.browserAction.enabled)
     }
 }
